+++ packages/component-library/src/components/icons-media/mt-icon/mt-icon.vue
@@ -108,7 +108,7 @@ watch(
 </script>
 
 <style lang="scss">
-@import "node_modules/@shopware-ag/meteor-icon-kit/icons/meteor-icon-kit.scss";
+@import "@shopware-ag/meteor-icon-kit/icons/meteor-icon-kit.scss";
 
 .mt-icon {
   display: inline-block;